修改导出函数中times的数据类型

+++ software/Hardware/PrinterMoto.h
@@ -5,6 +5,6 @@
 void PrinterMoto_Init(void);
 void PrinterMoto_Run(uint8_t motor);
 void PrinterMoto_Stop(void);
-void PrinterMoto_Run_Circle(uint8_t times);
+void PrinterMoto_Run_Circle(uint16_t times);
 
 #endif
